% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ggplot_to_ayab}
\alias{ggplot_to_ayab}
\title{Convert ggplot plot to image suitable for knitting with AYAB}
\usage{
ggplot_to_ayab(p, width = NULL, height = NULL, bw_method = NULL)
}
\arguments{
\item{p}{The ggplot plot}

\item{width}{The desired output width in pixels}

\item{height}{The desired output height in pixels}

\item{bw_method}{One of "threshold" or "quantize" (default).
There are two black and white conversion methods, because
how well the conversion works depends a bit on the properties of the image
given as input (especially colors). The two methods are "quantize" and
"threshold".}
}
\value{
A magick image object
}
\description{
Convert ggplot plot to image suitable for knitting with AYAB
}
