% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ggplot_to_magick}
\alias{ggplot_to_magick}
\title{Create magick image object from ggplot plot object}
\usage{
ggplot_to_magick(p)
}
\arguments{
\item{p}{The ggplot plot to convert}
}
\value{
A magick image object
}
\description{
Used internally.
}
