% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ggplot_void}
\alias{ggplot_void}
\title{Process ggplot object to remove everything except plot area}
\usage{
ggplot_void(p)
}
\arguments{
\item{p}{The ggplot plot to process}
}
\value{
A ggplot plot with only the plot area and everything else stripped
}
\description{
Process ggplot object to remove everything except plot area
}
