% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_test_plot}
\alias{make_test_plot}
\title{Make a simple black and whit grid test ggplot plot}
\usage{
make_test_plot()
}
\value{
A ggplot plot
}
\description{
Make a simple black and whit grid test ggplot plot
}
