% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_test_plot_complex_lines}
\alias{make_test_plot_complex_lines}
\title{Plot a timeline test ggplot plot with lines}
\usage{
make_test_plot_complex_lines()
}
\value{
A ggplot plot
}
\description{
Plot a timeline test ggplot plot with lines
}
