% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_text_img}
\alias{make_text_img}
\title{Create magick image object with text}
\usage{
make_text_img(input_text)
}
\arguments{
\item{input_text}{Character string to plot}
}
\description{
Create magick image object with text
}
