% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayab-package.R
\docType{package}
\name{rayab-package}
\alias{rayab}
\alias{rayab-package}
\title{rayab: R package for AYAB compatible images}
\description{
R package to create images from image or data in a format ready to be knit using the AYAB (all yarns are beautiful) software
}
\author{
\strong{Maintainer}: Mario Angst \email{mario.angst@uzh.ch}

}
\keyword{internal}
