% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{save_ayab_png}
\alias{save_ayab_png}
\title{Save magick image object to png}
\usage{
save_ayab_png(fig, filepath)
}
\arguments{
\item{fig}{The image object}

\item{filepath}{Desired output file path}
}
\value{
Saves object to file. Returns the file path.
}
\description{
Save magick image object to png
}
