% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{text_to_ayab}
\alias{text_to_ayab}
\title{Convert text to image suitable for knitting with AYAB}
\usage{
text_to_ayab(input_text, width = NULL, height = NULL)
}
\arguments{
\item{input_text}{The desired text to be knit}

\item{width}{The desired output width in pixels}

\item{height}{The desired output height in pixels}
}
\value{
A magick image object
}
\description{
Convert text to image suitable for knitting with AYAB
}
