% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayab-package.R
\docType{package}
\name{rayab-package}
\alias{rayab}
\alias{rayab-package}
\title{rayab: R package for AYAB compatible images}
\description{
Functionality to create images suitable for two-color knitting via the AYAB (all yarns are beautiful) open source software/ hardware project from plots, text and images.
}
\author{
\strong{Maintainer}: Mario Angst \email{mario.angst@uzh.ch}

}
\keyword{internal}
